feat: document boundary implementation refactor extract and classifcation pipeline
