+++ tests/integration/check_extract_pipeline.py
@@ -35,8 +35,8 @@ def run_extract_pipeline():
     # setup_storage()
     MDC.put("request_id", "test")
     img_path = "~/tmp/address-001.png"
-    img_path = "~/tmp/grapnel/aligned/PID_2_11058_0_201458362.tif"
-    img_path = "/tmp/generators/87d95063109ffb8d4bad5698e433700b/burst/PID_3558_10887_0_201774185_00001.tif"
+    img_path = "~/tmp/demo/159000487_1.png"
+    img_path = "~/tmp/4007/176075018.tif"
     img_path = os.path.expanduser(img_path)
     # StorageManager.mkdir("s3://marie")
